+++ src/main/java/com/invadermonky/magicultureintegrations/api/item/ItemBaubleMI.java
@@ -5,13 +5,14 @@ import baubles.api.IBauble;
 import baubles.api.cap.BaublesCapabilities;
 import baubles.api.cap.IBaublesItemHandler;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.*;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.EnumActionResult;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.EnumHand;
 import net.minecraft.world.World;
 import net.minecraftforge.items.ItemHandlerHelper;
-import vazkii.botania.common.core.helper.PlayerHelper;
 
 public abstract class ItemBaubleMI extends Item implements IBauble {
     @Override